Working on being able to run multiple listeners in the same application.
The idea is that once the application is started, one do call
iserve:add_server(Port, Callback, Arg) and a server is started
that listens on Port and calls Callback:request/3 (with Arg)
in the arguments. And that several of these servers can be started
all being managed by the iserve application.

The starter of the server should end up being linked to the server
process, so they clean up well on exit, crash-only style.

Also setting up build into test/ for some lacking eunit-tests.